requirements.txt
ytmusicapi>=1.3.0
anthropic>=0.40.0
python-dotenv>=1.0.0
flask>=3.0.0
